users/requirements.txt
nameko
nameko-redis
pyjwt